requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
multithreading==0.2.0
mysql-connector-python==8.0.24
numpy==1.20.2
pandas==1.2.4
protobuf==3.16.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
schedule==1.1.0
six==1.16.0
urllib3==1.26.4
websocket-client==0.59.0
wincertstore==0.2
